Refactor cmd/collector package (#337)
* Refactor collector cmd codebase

The cmd codebase was getting cluttered, and this aims to simplify and
restructure how collector startup occurred.

Testing Done: no functionality changed, tested manually.

* Add license to all new files
