+++ cmd/occollector/app/collector/healthcheck.go
@@ -1,0 +1,37 @@
+// Copyright 2019, OpenCensus Authors
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package collector
+
+import (
+	"flag"
+
+	"github.com/jaegertracing/jaeger/pkg/healthcheck"
+	"github.com/spf13/viper"
+	"go.uber.org/zap"
+)
+
+const (
+	healthCheckHTTPPort = "health-check-http-port"
+)
+
+func healthCheckFlags(flags *flag.FlagSet) {
+	flags.Uint(healthCheckHTTPPort, 13133, "Port on which to run the healthcheck http server.")
+}
+
+func newHealthCheck(v *viper.Viper, logger *zap.Logger) (*healthcheck.HealthCheck, error) {
+	return healthcheck.New(
+		healthcheck.Unavailable, healthcheck.Logger(logger),
+	).Serve(v.GetInt(healthCheckHTTPPort))
+}
