+++ cmd/occollector/app/collector/logger.go
@@ -1,0 +1,42 @@
+// Copyright 2019, OpenCensus Authors
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package collector
+
+import (
+	"flag"
+
+	"github.com/spf13/viper"
+	"go.uber.org/zap"
+	"go.uber.org/zap/zapcore"
+)
+
+const (
+	logLevelCfg = "log-level"
+)
+
+func loggerFlags(flags *flag.FlagSet) {
+	flags.String(logLevelCfg, "INFO", "Output level of logs (TRACE, DEBUG, INFO, WARN, ERROR, FATAL)")
+}
+
+func newLogger(v *viper.Viper) (*zap.Logger, error) {
+	var level zapcore.Level
+	err := (&level).UnmarshalText([]byte(v.GetString(logLevelCfg)))
+	if err != nil {
+		return nil, err
+	}
+	conf := zap.NewProductionConfig()
+	conf.Level.SetLevel(level)
+	return conf.Build()
+}
