+++ internal/collector/opencensus/receiver.go
@@ -17,55 +17,38 @@
 package ocreceiver
 
 import (
+	"context"
 	"fmt"
-	"net"
 	"strconv"
-	"time"
 
 	"github.com/spf13/viper"
 	"go.uber.org/zap"
 
-	agenttracepb "github.com/census-instrumentation/opencensus-proto/gen-go/agent/trace/v1"
 	"github.com/census-instrumentation/opencensus-service/cmd/occollector/app/builder"
-	"github.com/census-instrumentation/opencensus-service/internal"
 	"github.com/census-instrumentation/opencensus-service/internal/collector/processor"
-	"github.com/census-instrumentation/opencensus-service/receiver/opencensus/octrace"
+	"github.com/census-instrumentation/opencensus-service/receiver"
+	"github.com/census-instrumentation/opencensus-service/receiver/opencensus"
 )
 
-// Run starts the OpenCensus receiver endpoint.
-func Run(logger *zap.Logger, v *viper.Viper, spanProc processor.SpanProcessor) (func(), error) {
+// Start starts the OpenCensus receiver endpoint.
+func Start(logger *zap.Logger, v *viper.Viper, spanProc processor.SpanProcessor) (receiver.TraceReceiver, error) {
 	rOpts, err := builder.NewDefaultOpenCensusReceiverCfg().InitFromViper(v)
 	if err != nil {
 		return nil, err
 	}
 
-	grpcSrv := internal.GRPCServerWithObservabilityEnabled()
+	addr := ":" + strconv.FormatInt(int64(rOpts.Port), 10)
 
-	// Temporarily disabling the grpc metrics since they do not provide good data at this moment,
-	// See https://github.com/census-instrumentation/opencensus-service/issues/287
-	// if err := view.Register(ocgrpc.DefaultServerViews...); err != nil {
-	// 	return nil, fmt.Errorf("Failed to register ocgrpc.DefaultServerViews: %v", err)
-	// }
-
-	lis, err := net.Listen("tcp", ":"+strconv.FormatInt(int64(rOpts.Port), 10))
+	ocr, err := opencensus.New(addr)
 	if err != nil {
-		return nil, fmt.Errorf("Cannot bind tcp listener to address: %v", err)
+		return nil, fmt.Errorf("Failed to create the OpenCensus trace receiver: %v", err)
 	}
-
 	ss := processor.WrapWithSpanSink("oc", spanProc)
-	oci, err := octrace.New(ss, octrace.WithSpanBufferPeriod(1*time.Second))
-	if err != nil {
-		return nil, fmt.Errorf("Failed to create the OpenCensus receiver: %v", err)
+	if err := ocr.StartTraceReception(context.Background(), ss); err != nil {
+		return nil, fmt.Errorf("Cannot bind Opencensus receiver to address %q: %v", addr, err)
 	}
-
-	agenttracepb.RegisterTraceServiceServer(grpcSrv, oci)
-	go func() {
-		if err := grpcSrv.Serve(lis); err != nil {
-			logger.Error("OpenCensus gRPC shutdown", zap.Error(err))
-		}
-	}()
 
 	logger.Info("OpenCensus receiver is running.", zap.Int("port", rOpts.Port))
 
-	return grpcSrv.Stop, nil
+	return ocr, nil
 }
